requirements-dev.txt
pytest==6.2.5
boto